refactor: move DIdService do domain chore: remove old services folder
